Removed some compiler casting warnings
Former-commit-id: e4f1bb782327a8eac4e89d5cf0cf47416dfc76b2
Former-commit-id: 4ea33f663a605c7d9c9f3c0e24b626a35eb89f9e

+++ source/variables/doublevariable.cpp
@@ -327,7 +327,7 @@ namespace HuwInterpreter {
             {
                 return nullVariable;
             }
-            long long value = static_cast<long long>(static_cast<long long>(this->value) | variable->toInt());
+            long long value = static_cast<long long>(this->value) | variable->toInt();
             return std::move(std::make_shared<DoubleVariable>(passable, value));
         }
 
@@ -337,7 +337,7 @@ namespace HuwInterpreter {
             {
                 return nullVariable;
             }
-            long long value = static_cast<long long>(static_cast<long long>(this->value) ^ variable->toInt());
+            long long value = static_cast<long long>(this->value) ^ variable->toInt();
             return std::move(std::make_shared<DoubleVariable>(passable, value));
         }
